requirements/requirements-test.txt
py==1.4.31
pytest==2.9.2
pytest-django==2.9.1